Intitially define allowed Flex Page Layouts as all
This pertains to UDK8-143
The empty array allows all layouts to be used on a given
content type. Previously, the undefined empty array
caused a warning produced in Layout Per Node. The code in
Layout Per Node should also be fixed to prevent non-existent
arrays from producing a warning (though this bug is really
due to our specific original implementation of the .yml file.

At a later point, this logic will be changed to explicitly
define which layouts should be allowed, by default, so this
is only a temporary setting.