Target display fix for <summary> tag to Firefox only
In the current vendor/normalize.css that ships with core, the <summary>
tag is given 'display:block', which removes the caret in Firefox. This is
pending a core fix -- see https://www.drupal.org/project/drupal/issues/2821525
In the meantime, we set the display to "list-item", but only for Firefox,
as doing so in MS IE/Edge adds an additional bullet display.