Use relative path for archiveArtifacts command (#1064)
We noticed that tests were no longer pushing the .html page output
to the artifacts.

A suggestion at https://stackoverflow.com/a/59073926 indicated that
the path needed to be relative, and we verified that that did once again
push the .html page output to the build artifacts.

However, since the initial copying of those artifacts was only happening
in the `catch` statement (i.e., when a test failed), a successful run
would not output the artifacts.

This led to examining why we were using the try/catch logic in the
first place. Removing that logic resulted in seeing the PHPUnit test
fail, but Jenkins would not recognize that as a failure.

This potentially could be obviated by removing the --stop-on-failure
flag, which would then complete the full test suite and report a failure
but would take longer.

Therefore, we decided to leave the try/catch logic as is.