1096 - Save promo list uri within link element (#1100)
* Save promo list uri within link element

* 1098 - Conditionally prepare link options to accommodate existing link data (#1101)

* Conditionally prepare link options
Release 3.0.0-beta.5 added the ability to have link options on the
custom widgets. However, links created prior to 3.0.0-beta.5 will not
have an "options" array attribute in the database. Therefore, this code
change conditionally adds those values if present, and otherwise
defaults to a prepopulated array structure with empty values.

* Properly set title with override, then default, then uri